refactor: extract shared fragment logic to extensions to avoid duplicated code
